Bug Fix in updateRating(int rate) in Class ServiceProviderUser and 3 test cases
Fixed a bug in updateRating(int rate) in Class ServiceProviderUser, that
caused the method to output incorrect results when noRated > 2.
Modified the class so now it keeps track of the total rating stored in
variable private double totalRating.
Changed the type of noRated from double to int since it is always
integer.
Modified the method updateRating(int rate) so that it rounds the result
to 1 decimal places.
Added 3 test cases and tested the above changes.